refactor(store): extract board persistence into dedicated module
Move PersistedNode, PersistedScenario, PersistedSlice, PersistedBoundedContext,
PersistedState interfaces plus STORAGE_KEY, DEFAULT_BC_ID, DEFAULT_BC_NAME,
defaultBoundedContexts(), loadFromStorage(), and saveToStorage() from
useBoardStore.ts into boardPersistence.ts.

Remove unused imports from the store (DomainEventNode, CommandNode,
ReadModelNode, PolicyNode, UIScreenNode, VerticalSlice, Scenario,
BoundedContext) that were only needed by the extracted persistence code.
